pyproject.toml
[project]
name = "fondi"
version = "0.0.1"
authors = [
  { name="Valter Yde Daugberg", email="valter@hotmail.com" },
]
description = "A small rendering for LaTeX mathematics"
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Operating System :: OS Independent",
]
dependencies = [
  "pillow",
]

[project.urls]
Homepage = "https://github.com/valteryde/fondi"
Issues = "https://github.com/valteryde/fondi/issues"

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]